+++ CHANGELOG.md
@@ -47,8 +47,9 @@ On the other hand, this also allowed us to deal with further issues that contain
 - Deprecated functionality
   - Removed `default_tensor_type()` method.
   - Removed `small_text.integrations.pytorch.utils.labels.get_flattened_unique_labels()`.
-- Classification
-  - Removed legacy arguments for early stopping in `__init__` for KimCNN and TransformerBasedClassification. (Use `fit()` keyword arguments instead.) 
+  - Classification
+    - Removed early stopping legacy arguments in `__init__` for KimCNN and TransformerBasedClassification. (Use `fit()` keyword arguments instead.) 
+    - Removed model selection legacy argument in `TransformerBasedClassification.__init__()`.
 
 ---